requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
ajpy==0.0.4
aniso8601==8.0.0
apparmor==2.13.4
appdirs==1.4.4
argcomplete==1.8.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
backdoor-factory==0.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
binwalk==2.2.0
bleach==3.1.4
blinker==1.4
Brotli==1.0.7
cairocffi==0.9.0
capstone==3.0.5
Cerberus==1.3.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
configobj==5.0.6
crackmapexec==5.0.0.dev0
crcelk==1.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
dhcpig==0.0.0
dicttoxml==1.7.4
distlib==0.3.1
distro==1.4.0
distro-info==0.23
dnslib==0.9.12
dnspython==1.16.0
EasyProcess==0.2.5
ecdsa==0.15
entrypoints==0.3
evdev==1.3.0
filelock==3.0.12
flasgger==0.9.4
Flask==1.1.1
Flask-HTTPAuth==3.2.4
Flask-RESTful==0.3.8
future==0.18.2
fuzzywuzzy==0.18.0
GDAL==3.0.4
GeoIP==1.3.2
geoip2==2.9.0
gevent==1.4.0
goofile==1.6
google==2.0.2
gpg===1.13.1-unknown
greenlet==0.4.15
grequests==0.4.0
gssapi==1.6.1
h11==0.9.0
h2==3.2.0
hashID==3.1.4
hpack==3.0.0
html5lib==1.0.1
httplib2==0.14.0
hyperframe==5.2.0
idna==2.8
imageio==2.4.1
impacket==0.9.20
ipdb==0.12.2
IPy==1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==6.1.2
jupyter-core==4.6.3
kaitaistruct==0.8
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.9.1
KismetCaptureRtladsb==2019.10.1
KismetCaptureRtlamr==2019.10.1
kismetexternal==2019.12.1
kiwisolver==1.0.1
lazy-object-proxy==1.4.3
ldap3==2.5.1
ldapdomaindump==0.9.1
LibAppArmor==2.13.4
logutils==0.3.3
lxml==4.5.0
lz4==3.0.2+dfsg
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mate-menu==20.4.1
mate-tweak==20.4.0
matplotlib==3.1.2
maxminddb==1.4.1
mccabe==0.6.1
mechanize==0.4.5
mega.py==1.0.8
meld==3.20.2
mistune==0.8.4
mitmproxy==4.0.4
mock==3.0.5
msgpack==0.6.2
multidict==4.7.5
mutagen==1.44.0
mysqlclient==1.4.4
nassl==3.0.0
nbconvert==5.6.1
nbformat==5.0.5
netaddr==0.7.19
networkx==2.4
nose==1.3.7
notebook==6.0.3
ntlm-auth==1.1.0
numpy==1.17.4
nyx==2.1.0
olefile==0.46
onboard==1.4.1
onioncircuits==0.6
onionshare==2.2
packaging==20.3
pandocfilters==1.4.2
paramiko==2.6.0
parsero==0.82
parso==0.6.2
passlib==1.7.2
patator==0.7
pathlib==1.0.1
pbkdf2==1.3
pbr==5.4.3
pefile==2019.4.18
persepolis==3.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.1
plecost==1.1.2
plotly==4.4.1
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.4
psutil==5.6.7
psycopg2==2.8.5
pyasn1==0.4.2
pycairo==1.16.2
pycares==3.1.1
pycountry==19.8.18
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.8
pycryptodomex==3.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydot==1.4.1
pyenchant==2.0.0
Pygments==2.3.1
PyGObject==3.36.0
pygraphviz==1.5
pyinotify==0.9.6
pylint==2.5.3
pylnk3==0.2.1
pymongo==3.10.0
PyNaCl==1.3.0
pynvim==0.4.1
pyodbc==4.0.22
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyperclip==1.7.0
PyQt5==5.14.2
pyqtgraph==0.11.0rc0
PyRIC==0.1.6.4
pyrsistent==0.15.5
pyrtlsdr==0.2.91
pyserial==3.4
PySimpleSOAP==1.16.2
pysmbc==1.0.15.6
pysmi==0.3.2
pysnmp==4.4.6
PySocks==1.6.8
python-apt==1.9.10
python-dateutil==2.7.3
python-debian==0.1.37
python-debianbts==3.0.2
python-dotenv==0.14.0
python-Levenshtein==0.12.0
python-libnmap==0.7.0
python-libtorrent==1.2.5
python-magic==0.4.16
python-magic-ahupp==0.4.13
python-memcached==1.58
python-xlib==0.26
pytz==2019.3
pyusb==1.0.2
PyVirtualDisplay==0.2.1
pywerview==0.2.0
pywinrm==0.4.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QtPy==1.9.0
redis==3.3.11
reportbug==7.6.0
requests==2.23.0
requests-ntlm==1.1.0
retrying==1.3.3
rfcat==1.0.1
roguehostapd==1.1.2
rq==1.2.2
rtlsdr-scanner==1.3.2
ruamel.yaml==0.15.89
scapy==2.4.3
scipy==1.3.3
selenium==4.0.0a1
Send2Trash==1.5.0
setproctitle==1.1.10
shodan==1.23.0
simplejson==3.16.0
sip==4.19.22
six==1.14.0
smoke-zephyr==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
specan==0.0.0
splinter==0.13.0
SQLAlchemy==1.3.15
sslyze==3.0.1
stem==1.7.1
tabulate==0.8.2
tenacity==5.1.5
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
termineter==1.0.4
testpath==0.4.4
texttable==1.6.2
theHarvester==3.1.0
tld==0.11.11
tls-parser==1.2.1
toml==0.10.1
torbrowser-launcher==0.3.2
tornado==5.1.1
traitlets==4.3.3
typing-extensions==3.7.4.2
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.25.8
urwid==2.0.1
vinetto==0.8.0
virtualenv==20.0.29
visvis==1.11.1
wafw00f==2.1.0
wapiti3==3.0.3
wcwidth==0.1.8
webencodings==0.5.1
websploit==4.0.4
Werkzeug==1.0.0
wfuzz==2.4.5
widgetsnbextension==2.0.0
wifiphisher==1.4
wifite==2.5.2
wrapt==1.12.1
wsproto==0.15.0
wxPython==4.0.7
xcffib==0.8.1
XlsxWriter==1.1.2
xmltodict==0.12.0
yara-python==3.10.0
yarl==1.4.2
yaswfp==0.9.3
youtube-dl==2020.3.24
